Complete implementation of smarter Induct_on
This massages a goal to pull out a negative occurrence of a inductive
relation term, replacing compound terms with variables and
generalising appropriately to make the goal suitable for a subsequent
call to ho_match_mp_tac (which is done inside Induct_on).

Not sure yet quite what should be done with mutually recursive
relations and Induct_on.

Closes #244