Rework handling of string literals
Can now do the parsing side of github issue #510, allowing string
literals to be invisibly injected into other types, just as numbers
are. Add to this feature with the ability to use the «...» delimiters
as a separate flavour of string (with the underlying term being the
application of an injection function to a "normal" string literal).

Pretty-printing still needs to be done.