Change INDUCT_THEN to cope with more induction theorem shapes
In particular, it previously assumed that outermost universal
variables in assumptions that were implications were always of the
same type as the induction variable. E.g., in a binary tree induction
axiom, the branch IH would be

   !t1 t2. P t1 /\ P t2 ==> !a. P (Nd a t1 t2)

(note how the !a is under the implication).

The magic done by INDUCT_THEN would then change the variables t1 and
t2 to match the name of the induction variable in the actual goal (one
would be primed). But if you have a snazzier principle with an IH such
as

  !ts. (!t. MEM t ts ==> P t) ==> !fnm. P (Fn fnm ts)

the outermost variable (which has to be outermost), is not of the
"right" type.

The fix allows for this, and also copes if the fnm quantification
floats out to the top-level.  This facility is only relevant if a
non-standard induction theorem is installed for a type (see previous
commit). The standard theorems generated by Datatype are shaped as
INDUCT_THEN required.