Extend and move theory of lists-of-nats encoded as nats
Previously, there was a rudimentary script in examples/computability's
prnlistTheory, which was used only in the computability tree. It's
useful to have all this stuff more widely available. (For example, in
showing Lowenheim-Skolem, setting up a bijection between the nats and
formulas is needed.)

New theory is now src/list/src/nlistTheory.

Thanks to Yiming Xu and Elliot Catt for help developing this theory.