Make some PREIMAGE theorems automatic rewrites
Also remove |- SING s ==> FINITE s from the set of automatic rewrites.

The presence of this is bad because a "Once" simplification theorem
meant for s may fire when the simplifier attempts to discharge the
hypothesis, and so it won't get to fire when actually rewriting s in
the "main" goal. (And no-one uses SING.)

Knock-on effects simplify folCanonScript.sml, the experience of
writing which prompted these changes in the first place.