Modify EmitTeX to allow \sp to stack properly after variable names
If you end up with \HOLFreeVar{foo\sb{\mathrm{1}}\sp{W} the
superscript W looks bad because the HOLFreeVar puts a \mathit around
its argument, and this interferes with the combination.  Now the map
is to

  \HOLFreeVar{foo}\sb{\mathrm{1}}

so that a \sp{W} after this stacks nicely.

In raw LaTeX, consider the difference between

  \mathit{foo_2}^W   % yuck

and

  \mathit{foo}_2^W   % nice