Adjust ARM/v7/arm_step tech in light of 44aabdd82
Along the way, make this SML code more independent of the stateful
rewriter; previously the code in arm_stepLib could change its
behaviour from one invocation to the next if the stateful simpset
changed. That's no longer true, but there's currently still a
dependency on the state of it when arm_stepLib is loaded.