Explicitly pass --holstate to Holmake when calling it in some tests
There is a new behaviour to (poly) Holmake here: all targets now
depend on the specified HOLHEAP, not just those that are just HOLish.
If not otherwise specified, this defaults to bin/hol.state, which is
fine once HOL is built. The tests adjusted here are run before that is
built, and so need to explicitly depend on something else.  Using
--poly_not_hol avoids having a dependency at all, and it's also
possible (as here) to specify the early hol.state0.