Change boss/Holmakefile so it doesn't build hol.state from hol.state
HOLHEAP = hol.state means that other things in this directory depend
on hol.state, including bossLib.uo.  But building hol.state requires
bossLib.uo. This is a loop.

Changing HOLHEAP to point at the numheap in src/num/termination more
accurately captures what was going on, but does mean that the selftest
in this directory will run on that heap (rather than on hol.state).