Fix various Holmakefiles to make a dependency explicit
It is now not enough to refer to files in other directories and assume
that the presence of those directories in the INCLUDES line will
ensure that the dependencies will get built.  (They will do if you use
--recursive-build, I suppose.)

Instead, it's better practice to make it clear that there is a
dependency by listing it as such. This was better practice before of
course; the Holmakefiles as written wouldn't cause rebuilds with a
change to the dependency (a heap in this case).