Holmake: key monitor function on combination of PID and tag-string
The latter was probably unique when the tags were the filenames that
were either from the current directory, or had paths attached.  But
if the printing form is the tag, which means the path is always
stripped, multiple files of the same name may be being built in
parallel. Then they can be distinguished by the PID of the process
that is building them.