Eta-expand GEN_REWRITE_CONV so that bounded rewrites work better
Things like

   Once th

generate a reference when add_rewrites is called, so add_rewrites
should only be called when there is a term argument to be applied to
rather than in advance, which has the effect of sharing the
exhaustible reference between multiple arguments.

Thanks to Johannes Åman Pohjola for the bug report.