Link fixedPointTheory to earlier (more general) posetTheory
Could dispense with some of the proofs entirely and just use what's in
posetTheory except that the less general proofs make it clear (by
inspection of the definitions) that lfp and gfp on sets are formed by
using BIGINTER and BIGUNION respectively. This info is hidden by the
generic proofs in posetTheory, which just give you that the fix-points
exist (internally, they are constructed by appeal to the glb and lub
known to exist in the complete po).