emacs-mode: Tweak handling of backticks once more
Hitting backtick in a holscript file should now be more convenient to
use.  I will document the exact behaviour in the interaction guide and
probably also the release notes.

Importantly, to get

  ‘foo ∧ bar’

the mode allows you to press the following keystrokes in exactly this
sequence

  `foo /\ bar`

Then, to get

  “foo ∨ bar ⇒ baz”

you type

  ``foo \/ bar ==> baz`

The visual feedback should make it clear what's happening when you do
this (and why you only need to close with one backtick).

[Note that the handling of \/, /\ and ==> becoming ∨, ∧ and ⇒ is a
 feature of hol-input, not the new handling of backticks.]