More ℝ-normalisation, removing inverted factors where possible
Previously, the code would collect up

    ...x...  <=   ... x...

for various possible exponents of x, but the normalisation would only
trigger if the x was present on both sides.  Now, if the exponent is
negative, a lone x will shift to the other side of the relation
regardless.

This applies also to fractions, so that

   x <= 3/4 * y

will turn into

   4 * x <= 3 * y

Fixes in src/probability and more explicit regression tests.