Fix decompilation of JAL instruction on RISC-V
The problem was that the decompiler treated JAL as a tail-call. It did
this because the decompiler thought the convention is to pass the
return address in register 14 (as is done on ARM). Because the JAL
instruction doesn't assign to register 14, it concluded the current
function's return address will be the called functions return address.

The solution is simply to make the decompiler aware that, on RISC-V,
the return address is passed in register 1. With this tweak the
decompiler correctly identifies JAL as a returning-call instruction.