Attempt at better Abbrev protection
This in the face of simplification and the variable elimination done
by RW_TAC and friends.  Core sequence, -t1 and some of -t2 builds (up
to category theory example).

Machinery for dynamically reverting to the old behaviours possible
through diminish_srw_ss and a trace variable.

This is work on github issue #800