emacs-mode: make parsing better aware of inner HOL syntax
Given,

   !x y z. p /\
   q ==>
      r

Hitting TAB on each line in succession (or selecting the whole block
and hitting TAB once) results in

   !x y z. p /\
           q ==>
           r

It won't introduce or remove newlines but will group things logically.

If you put a newline after the ., you get

   !x y z.
     p /\
     q ==>
     r

It should also be possible to make it aware of let and if-then-else
syntax.