mc/graph: use `r10` for `ret_addr_input` on RISC-V
The `ret_addr_input` parameter in function signatures produced by the
machine-code-to-graph decompiler is intended to receive the address of
memory that a caller expects a callee to use for return values that
won't fit into the usual return-value registers.

In the RISC-V calling convention, the `a0` (a.k.a. `x10`) register is
used for this same purpose. This commit ensures that `ret_addr_input` is
always passed the current value of `x10` (`r10` in the decompiler).

The `ret_addr_input` parameter is used by the seL4 [graph-refine][]
toolchain to make it easier to state certain properties relating to
function calls. It is not otherwise used in the function graphs produced
by the decompiler, so this change does not affect the proof of
correctness for the decompiler.

[graph-refine]: https://github.com/sel4/graph-refine