emacs-mode: Allow _ to be "symbol" or "word" in Emacs syntax tables
This can be controlled by user-preference:

  (modify-syntax-entry ?_ "w" holscript-mode-syntax-table) ; word
  (modify-syntax-entry ?_ "_" holscript-mode-syntax-table) ; symbol

The mode will continue to work with either choice. If you use
word-movement/manipulation commands (M-b, M-f, M-t, etc) the choice
here will cause emacs to see something like foo_bar as two words or
just one.

For the moment, the default is to have the first option (_ as 'word').

The regression tests are run with each option enabled.