emacs-mode: don't make _ a word character in syntax tables
Reading resources like

   https://emacs.stackexchange.com/questions/1075/whats-the-difference-between-words-and-symbols/

tells me that fiddling with the "w" class is bad form, and there is
this nice superword-mode that people can enable if they really want
M-f to jump over foo_bar as one word instead of two.

I do stick with ' as a word constituent though, handling the common
case where we have prime characters at the end of variable names.  If
we were more inclined to use it as a separator (e.g., foo'bar), then
this would probably also be worth reconsidering.