Add ind_for (:thm->thm) to DefnBase
This ind_for function is a cleaned up version of a function that
has existed for sometime in the CakeML translator. I'm moving the
function into DefnBase because it can be of more general use.

Example use:

    Datatype:
      prog = Assign num num | Seq (prog list)
    End

    Definition eval_def:
      eval s (Assign n v) = s⦇n ↦ v⦈ ∧
      eval s (Seq progs) = eval_list s progs ∧
      eval_list s [] = s ∧
      eval_list s (p::ps) = eval_list (eval s p) ps
    End

The definition of eval does not produce an induction theorem.
However, in proofs about eval it is often desirable to have
an induction theorem that is tailor to the types and exact
arguments used by eval.

With the new addition to DefnBase, one can get hold of such
an induction theorem as follows:

    Theorem eval_ind = DefnBase.ind_for eval_def