Define an adjacent relation on lists and their elements
So: adjacent L a b is true when a and b are adjacent (in that order)
within list L.

This allows a cute characterisation of what it is to be sorted.  Along
the way, make SORTED_DEF an automatic rewrite rather than just the
first two clauses of that definition, and then fix the fallout.