Fix bug in PAT_ASSUM when use of variables from goal is not respected
In particular, the old code gave itself licence to instantiate type
variables in patterns when those were associated with variables from
the goal, thereby allowing it to pick completely different
assumptions that didn't even include the quoted variable. The test
case illustrates.

Where a proof breaks (as in io_onestepTheory), the fix is to remove
the use of goal-variables of the wrong type, perhaps replacing them
with underscores.