Make Reference manual letter sections begin with correct entry
In particular, it previously began the sections with the first occurrence
of a lower-case letter; now it begins correctly with the first
occurrence, whether upper or lower case.

Also rename files corresponding to some symbolic identifiers to simply
have those names on disk too. Some characters (>, &) require
backslashes when typing these names in typical Unix shells, but as
long as we don't want to document /, we might as well use the
characters we're allowed to.