Add a new set of "global simplification" tactics
(Thanks to James Shaker for the suggested name.)

The "global" means that these simplify every assumption with respect
to every other assumption before eventually simplifying the goal as
well.

The core primitive is in simpLib and takes a simpset and a
configuration record as a parameter. The bossLib variants use letters
in the name to encode the common configuration options, and always use
the standard simp/fs/rw stateful simpset.  These are gs, gvs, gns and gnvs.

The addition of the 'n' and the 'v' toggle independent behaviours:

 'n'  -  don't use 'strip_assume_tac', meaning that goals won't
         strip.  (Existential variables won't get chosen and conjuncts
         won't strip either; I may go back on both of these.)
 'v'  -  eliminate variables (like VAR_EQ_TAC)

I hope that gs (or gvs) will replace the need to do the

  fs[] >> rfs[] >> rveq >> fs[]

"dance" in complicated proofs with big assumption lists.