Avoid combining star for RTC operator; use Slavonic Asterisk instead
The meaning of "combining" in the combining star character (U+20F0) is
that it should combine with previous characters in the same way as
separate accent characters (such as acutes: é) can combine with
predecessors. Using it as a pseudo-superscript just because it looks
OK in my particular Emacs handling of fonts and characters is a bad
idea.  In other contexts, I have seen it actually combine, and it
looks terrible.

So, instead of trying to solve the problem of general superscripting,
I'm changing to U+A673, which is a whole character in itself (not a
combining accent), is an asterisk, and seems to be generally typeset
relatively high on the line.  It's "open" which is a bit strange, but
otherwise seems to look OK.  Here's an example:

  RTC_SUBSET: ⊢ R x y ⇒ R꙳ x y