Prove (finite) converse of Kraft inequality
If you have a list of desired code-lengths and they "fit", then there
is indeed a prefix-free encoding that maps from codes of the desired
lengths into the indexes of the given list.  (Those indexes could then
indicate the "real" objects of interest.)

E.g., if you want a code for A .. Z, you provide 26 desired lengths
such that Σ 2^{-len(i)} ≤ 1 (e.g., they could all get length 5), and
there will be a prefix-free code for those letters. (In this case, a
bunch of 5-digit binary numbers.)

In the same example, there's space to allow 6 letters to have widths
of 4, and the rest with width 5.  (6/16 + 20/32 = 1) E.g.,

   A   0000   G  01100   N  10011   U  11010
   B   0001   H  01101   O  10100   V  11011
   C   0010   I  01110   P  10101   W  11100
   D   0011   J  01111   Q  10110   X  11101
   E   0100   K  10000   R  10111   Y  11110
   F   0101   L  10001   S  11000   Z  11111
              M  10010   T  11001

The buildL constant builds the binary tree encoding this mapping, with
an extra accumulator parameter than can be started at 0.