Make a use of raw_backend utility actually do something
It turns out that print_from_grammars sets up the backend that will be
used when printing is being performed.  Therefore the
testutils.raw_backend utility needs to be applied like so:

    raw_backend print_from_grammars min_grammars

in order for the reference to get changed appropriately.

This selftest worked because it was run in a situation where the
backend was already raw_terminal.  After the construction of the
heaps, the backend becomes something "interactive" by default.

Also make other tests use raw_backend so that this selftest file can
be run inside a hol.bare or hol environment.