Again allow ancestral data to consume deltas all at once on theory-load
This allows named simpset-fragments to be added to srw_ss.  This then
allows those fragments to be easily removed with calls to
diminish_srw_ss.  In the 'fix-ancestry' rework this ability was
removed but this did break proofs when diminish_srw_ss calls had no
effect.

Selftest in coretypes demonstrates this with removal of the "one"
fragment.