Tweak p/printer to sometimes put quantifiers next to large bodies
In particular, boolTheory.ABS_REP_THM changes from

  ⊢ ∀P.
      (∃rep. TYPE_DEFINITION P rep) ⇒
      ∃rep abs. (∀a. abs (rep a) = a) ∧ ∀r. P r ⇔ rep (abs r) = r

to

  ⊢ ∀P. (∃rep. TYPE_DEFINITION P rep) ⇒
        ∃rep abs. (∀a. abs (rep a) = a) ∧ ∀r. P r ⇔ rep (abs r) = r

At the moment, it will only do this if there is exactly one bound
variable of length/size 1.  The larger the quantified variable
blocks, the less horizontal space the body will get.  I think
saving a line for two characters lost in width is worth it.