Add `SELECT_GOALS_LT` (with `>>~` combinator) and `SELECT_GOALS_LT_THEN`
The former is similar to `SELECT_GOAL_LT`, but can select multiple goals. The latter additionally allows users to supply an additional tactic, the success of which will be used as an additional goal selection criterion. The selected goals will have this tactic applied to them in the resulting goal state.