Generalise quantifier-block-smashing behaviour from ef834d929
Now there is a trace called "PP.qblock_smash_limit". The larger the
number this has, the larger the block of leading quantifications can
be and still have the body appear on the same line. The number is the
indentation level, so the theorem below is using a limit of 10 (any
larger number would also serve):

   ⊢ (∀p m x. x ∈ p ⇒ RES_ABSTRACT p m x = m x) ∧
     ∀p m1 m2. (∀x. x ∈ p ⇒ m1 x = m2 x) ⇒
               RES_ABSTRACT p m1 = RES_ABSTRACT p m2

Putting this under a trace means that p/printing is easy to adjust in
the setting of a paper because the \HOLthm functions can tweak trace
variables on the fly. One might write

   \HOLthm[tr'PP.qblock_smash_limit'=10]{bool.RES_ABSTRACT_DEF}

to get something like the above, for example.