Close file-descriptors properly in (poly) Holmake
Issue was that file-descriptors connecting the main process to the
children actually doing the work were not being closed in the parent
even after the children had exited. If the OS's max-open-fds number
was only in the hundreds (on MacOS, it's 256 by default), you
could relatively easily see a

   Exception: SysErr ("Too many open files", SOME EMFILE)

type message (in big CakeML builds, for example). In older versions of
Poly/ML garbage collection would close these file-descriptors for us
as the relevant pieces of data expired, but no longer!