Begin new approach recording rightmost error for successful parses
This is important for things like repetition-forms.  For example, if
you have a FOO non-terminal and attempt to parse (rpt FOO) in a
context where there is a line of input which looks like a FOO right up
to character 60 and then causes the FOO to fail.  The current approach
would have the rpt succeed at the start of the input with nothing
consumed.  This then leads to very unhelpful error-reporting.

Committing because I am going to go back and get rid of locs; I need
these to be monotonically increasing so I now want the peg engine to
track these numbers/locations itself.