Remove minisat's makefile's generation of dependency info
Replace the dynamically generated info with a static copy of latest
version. Given that there are only four source (.C) files in this
directory, the complexity of just putting these directly into the
Makefile doesn't seem such an imposition.

We don't expect to ever be changing this source code in significant
ways either.