Extend "big product" construction in cardinalScript
In particular, realise that you may want to take a product over an
indexed set of sets, where there may be duplicate sets in the range of
the index function.  If you only have BIGPROD taking a set of sets,
you can't have two copies of the same set. BIGPROD can be defined in
terms of the indexed BIGPRODi because you can take the identity
function as the index function (modulo need to use option types for
partiality).

Add theorem that as long as neither side has a "zero term" (the empty
set), then products are injective.