Demonstrate the BNF approach on a concrete example (no axioms!)
The example is chosen to be moderately complicated, such that

  newty = C1 (num |-> newty) | C2 num (('a # newty) list)

Nice induction and recursion principles drop out pretty cleanly.

This would be a reasonable test-case for an implementation. To make it
trickier, you could test making the domain of the finite map be
'a. (And then handle nested and mutual recursion of course.)
Automating most of the proofs looks straightforward enough. The logic
to build the right definitions is complicated (building compositions
of functors is indeed one of the BNF approach's major contributions).