Add support for passing variable defs to Holmake
Adds support for VAR=value style commandline arguments to Holmake.
After parsing commandline options, the remaining arguments are separated
into those that match "ANYTHING=ANYTHING" and those that do not. The
latter are treated as before, while the extra variable definitions are
inserted into the base environment /after/ the holpathdb is read, and
/before/ the Holmakefile is read, each time Holmake reads a Holmakefile
(this is the intent, at least).