Parallelise build of core HOL
Still not quite in an ideal state, but close.  There are effectively
now two copies of the build "sequence" for the later HOL theories; one
in the sequences files, and the other in the Holmakefile in
parallel_builds/core.

This should be fixable by making it easier for Holmake to link things
into sigobj as part of its directory building. It would also be nice
to parallelise with examples building. I hope this will be as simple
as adding a parallel_builds/examples1/Holmakefile that refers to the
core/Holmakefile as well as the relevant examples directories.