Change build so that it no longer automatically calls selftest.exe
Instead, arranging for this to happen is the responsibility of the
Holmakefile in the relevant directory.  It's helped in this way by the
knowledge that build will define the HOLSELFTESTLEVEL variable to
equal the test level. Thus, an ifdef HOLSELFTESTLEVEL can be used to
have test programs called or not. Use this facility to have some
selftests generate .log files that record the running of a file.

This then saves on having the selftest rerun if nothing relevant has
changed.

A number of selftest.sml scripts need to be enabled for this new world
under src, as well as a few in examples.