Prove equivalence (with caveats) for `BranchImmediate` calls
The caveats are:
 - restrictions on some ASL system registers which do not seem to be modelled in L3
 - an attempt to remove a feature from the ASL model which does not seem to be modelled in L3

For the former, the system registers involved are `SCR_EL3` and `HCR_EL2`.

For the latter, the feature appears to be `FEAT_PAuth`, to do with pointer
authentication codes (PACs) and made mandatory in Armv8.3 (i.e. a later version
than that of the L3 model).
The workaround introduces the assumption `¬HavePACExt ()` into the proof,
but really the converse holds.