Make Holmake output more informative
It is not yet as pretty as I'd like, but it dispenses with one piece
of time info and prints out the directory of the Holmakefile with the
just completed target instead. When doing big multi-directory builds
it can be useful to see where the various things are being built are
coming from.