Holmake: "link-to-sigobj" now can/should be done in Holmakefiles
bin/build sets variable HOLBUILD to 1 when it calls Holmake, so
Holmakefiles can now arrange to move stuff into sigobj themselves. The
tricksy shell-command to do this is available through the built-in
$(HOL_LNSIGOBJ) variable.

More directories mentioned in src/parallel_builds/core's Holmakefile
need to have this adjustment made.

For the moment, bin/build will continue to do the upload to sigobj as
well when it stops working in non-test directories from its sequence.