Recurse to build subtype combinator eqs
In an unusual case where a type recurses to itself via a
datatype which itself has auxiliary size constants, the
normal forms for those constants are required.

These could be saved and fetched by changing the TypeBase
interface, but, it isn't really worth it, since the
constructors of the inner type's recursion are all involved
in the current theorem/induction anyway, so the ratio of
duplicated work is tiny.