Split sum_size to match default shape
As with prior patches to option_size and pair_size, sum_size
in basicSizeTheory should have the same shape as DataSize.define_size
would have given it. However, sum_size is actually in use in defining
measures, where a +1 would be a problem.

Instead, add a new full_sum_size with the correct shape, and register
it instead of sum_size with TypeBase.