Remove extrealTheory.pow_zero_imp as automatic rewrite
This theorem states

  (∃n. x pow n = 0) ⇒ x = 0

This is an extremely bad rewrite: the lhs of the equation (x) matches
every possible extreal term, and then causes the simplifier to try to
solve

  ∃n. x pow n = 0

Not only is the simplifier typically poor at solving existentials, the
body under the quantifiers matches exactly the same rewrite all over
again.

Thanks to Jared Yeager for helping me find the problem.