Holmake: change dependency analysis to prefer INCLUDES over sigobj
This should solve the issue observed in github pull request #965,
where a rebuild could try to load a file from sigobj without that file
actually being there.  The trick is to make sure that dependencies are
never listed as being in sigobj if they can be found elsewhere.  This
then ensures that the parallel build approach to building the core
system prefers to record dependencies in src directories rather than
sigobj, even if they appear to be in both (given that the latter is
just a symbolic link to the former).

This makes the treatment of INCLUDES almost identical to that of
PRE_INCLUDES, but it's possible there's some utility in being sure
that directories in the latter will be searched before those in the
former, so I'm keeping PRE_INCLUDES for the moment.

Update the documentation and add a test-case demonstrating that an
INCLUDES spec can mask the built-in tautLib.  (Note that things
embodied in a heap can't be masked, so the test-case won't work if
moved to later in the build sequence when the core Poly heap already
has tautLib preloaded.)