Give user-congruences in simplifier "UNCHANGED" optimisation
In particular, avoid having to do an aconv check after a congruence
has finished working on its sub-terms.  In order to get this all to
make sense, also modify the behaviour of "higher-order"
congruences such as MAP_CONG:

  l1 = l2 ==>
  (!e. MEM e l2 ==> f1 e = f2 e) ==>
  MAP f1 l1 = MAP f2 l2

so that they avoid introducing eta-redexes.  This latter change
requires some proof mangling in places where heavy use of MAP_CONG was
made alongside TFL definitions that used MAP and introduced
eta-redexes of their own.  (This is probably an argument to have TFL
automatically post-process to remove those.)