Holmake: run hollogs test only if -j option > 1
A bin/build -j1 -t under Poly would fail in this test otherwise. (The
build sequence already ensures that a Moscow ML build doesn't try to
build this directory.)

Enable a check for whether or not this condition is true by adding
HOL_NUMJOBS variable to those that Holmakefiles can refer to.