Stop boss simp routines (fs,simp,...) repeatedly rebuilding simpsets
These tactics are based on srw_ss but merge in ARITH_ss and LET_ss
each time they're called. Better to have them do this only when
necessary (i.e., when the underlying srw_ss() value has changed).
Generalise the idea behind this to allow arbitrary values to be based
on the srw_ss simpset.