Move TotalDefn.termination_simps to ThmSetData
This adds "tfl_termsimp" as a new theorem attribute.

Other extra facilities are then easy to provide on top of this,
including temporary adjustments to the set, which is illustrated in
the n-bit selftest.

There is some discussion on github pull request #975.

Still to do: the same change for the WF and termination-solve-simp
theorem lists.