Move definition of root and sqrt from transc to real
Given these are the essence of algebraic numbers it seems particularly
inappropriate to have them defined in transc.  The properties proved
of them can stay in transc, but having the constants here mean that
one can refer to sqrt (as done in binary_ieee) without needing to suck
in a whole lot of topology-based material.