Change iterate's _ .. _ notation to { _ .. _ }
In the grammar, the .. operator is no longer an infix, but part of a
"close-fix" notation involving the braces.  This is the same as the way
listRange sets up its [ m .. n ] syntax.

Most (but not all) uses end up needing to parenthesise, e.g., in
things like

  sum (m .. n)

so changing to braces doesn't waste characters.

Note that this notion has nothing to do with real numbers, so I will
work towards moving it into pred_set at some future date.  There are
also lots of theorems about it that should be automatic.