Update INCLUDES info for Hoare-for-divergence example
Problems like this can be masked by non-determinism of build order; if
in a big parallel build, the directories the example depends on are
fortuitously built and uploaded to sigobj before this directory
begins its execution, then the error is not seen.

Also, of course, if all of the standard build is done, then the
examples will build if Holmake is invoked in their directories
subsequently.