Fix INCLUDES for src/coalgebras
There may yet be errors of this sort lurking. The build process has
become somewhat non-deterministic thanks to parallelisation: if every
previous build has happened to build a dependency and put it into
sigobj before the directory depending on it has started, things may
appear to be fine when, strictly speaking, they are not.