vimhol: enables hosting the hol repl inside vim
This patch enables to open a hol repl in a terminal buffer inside the
vim editor, and introduces keyboard shortcuts for interaction. These
changes only have an effect for vim with the 'terminal' feature (if
`:echo has('terminal')` shows `1`), and for neovim.

The shortcut `hx` opens a new hol repl in the current working directory
of the active buffer (i.e. file). The usual shortcuts can be used for
interaction with the most recently opened hol repl.
Pressing `hX` terminates the hol repl. (It sends `ctrl-d` to hol).

Non-standard heaps do not load the code to setup the vim-hol repl
connection. To help, set the `$HOL_CONFIG` environment variable, for
example to "$HOLDIR/tools/vim/hol-config.sml". The new shorthand `hC`
will load this config file which connects vim and the hol repl.

In addition `hw` types the current selection or line in the (neo)vim
terminal buffer.