Store induction theorems for constants when they are defined
This only happens if the qDefine entry-point and/or the Definition
syntax is used.  It also only stores tfl induction principles, not
those for types where the definition could have been made by primitive
recursion.  This decision is certainly open to revision.

Modify the types of TotalDefn.{t,x}Define to also return induction
theorems (as above), breaking code that was using either. Being forced
to update broken code is good because it then ensures that the
induction theorem for the definition will be stored in DefnBase.

Naturally, a number of examples are still broken.  This is the bulk of
what needs to be done for github issue #627.