Reinstate the old xDefine and tDefine functions (at least in bossLib)
I realised that they had obvious implementations in terms of qDefine,
and this should make life easier (less in the way of backwards
incompatibility), and still gets the advantage of DefnBase goodness.

As for all those other fixes I made earlier, the Definition syntax is
still nicer: I regret nothing!