Tweak one_line_ify to handle vacuous patterns
A vacuous pattern is something like

  f (x,y) = 3

where the arguments under the constructor don't contribute to the
result, and the type only has one constructor. Every use of () (unit) is a
vacuous constructor.  Arguably the user is just making their own life
difficult by writing such definitions, but hey.

Some vacuous patterns can be entirely eliminated with the code now,
but others persist, as in the nested vacuous test case, which takes

   nestedv f a (INL p) = f p /\
   nestedv f a (INR (x,y,z)) = f (a, F)

to

   nestedv f a v = case v of INL p => f p | INR (x,y,z) => f (a,F)

which is correct, and on one line, but still dumb.