Open listTheory in rich_listScript
This was avoided for the longest time, and for reasons that are lost
to history.  Things are a lot simpler if we just open listTheory and
be done with it.  The only issue arising is the way the rich_list
script wants to refer to something called REVERSE.  This is specially
bound at the head of the file.