Move datatype size_eq self-tests into a theory
It's important to test size_eqs with the basic types (with their
hand-rolled size functions) to make sure everything is working.
This is moved into a separate theory that explicitly depends on
basicSize, since in the current setup selftest does not seem to
pull in basicSize (which may or may not be surprising).