vimhol: set temporary pipe properly for macOS
The vimhol convenience script used `mktemp -p` to create a new temporary
pipe for sending commands to an instance of hol. According to GNU
coreutils `-p` followed by a directory is supposed to create said pipe
relative to a given directory. As reported this seems not available on
macOS.
The behaviour of `-p` is emulated by setting a `$TMPDIR` environment
variable.