vimhol: enable handling multiple hol sessions
New hol sessions (started via `hx`) in a vim terminal are pushed on a stack of
sessions. Commands to hol are send to the most recently opened hol session,
i.e. the top of the stack. Whenever a session is closed via `hX` it is popped,
and the previous session becomes active. This is very useful when working with
multiple files in different theory contexts.
To keep an overview of the stacked sessions, the vim terminal shows the file
name from where the session was started.

(These changes also avoid some cases of sending a command to a non-existing
pipe.)