Add itree definition with Tau
The type name is itree, which coincides with the type
name from itreeTheory. I decided to keep the itree name
because it's the most natural name and most uses will
only load one of the two versions. Files that load both
ought to install an type abbrev to see the difference.