Use modern syntax a deal more in binary_ieeeScript.sml
This has the big advantage in this case of inserting semi-colons
between theorems, meaning that the REPL's context is generally
correct when an exception is raised.  If a series of declarations is
made without semi-colons and the last raises an exception, Poly/ML
and Moscow ML (at least) won't put any of them into the SML
name-space.