Force the CBV_CONV conversions to lock their compsets before proceeding
This is because both call initial_rws, and that function in turn calls
clauses.from_term, which calls clauses.assoc_clause, which makes a
reference update for an entire tree corresponding to the compset.  We
can't afford to have such calls interleave because only one update may
survive.

The result is almost no calls to rwlock.read in all of the computeLib
API.