vimhol: honouring lastmaker
The vimhol plugin (with (n)vim terminal feature) now dynamically chooses
the first hol executable in the following order, when starting a new
interactive hol session:

1. hol relative to the lastmaker
  If `Holmake` was run on the edited theory before, its path resides in
  `.HOLMK/lastmaker`. vimhol probes that path with `hol` substituted for
  `Holmake`.
2. `$HOLDIR` environment variable
3. `hol` if in scope, e.g. from `PATH`

This feature removes the earlier requirement of setting a `$HOLDIR`
environment variable.

The path of the chosen hol is printed as a message for the user (via
`echomsg`), which is as well accessible through `:messages`. The message
has the shape `hol repl cmd(<id>): <cmd>` containing the related
terminal buffer id and the hol command.